requirements.txt
httpx>=0.27
lxml>=5.0
openai>=1.0
pyyaml>=6.0
